unit test for replay parsing (#1744)
* add parsing test

* just directly push to result array

* generate path to find test file
