+++ src/extension/replay/vscode-node/replayDebugSession.ts
@@ -18,6 +18,7 @@ import * as fs from 'node:fs';
 import * as path from 'node:path';
 import { commands, type WorkspaceFolder } from 'vscode';
 import { ChatReplayResponses, ChatStep } from '../common/chatReplayResponses';
+import { parseReplay } from '../node/replayParser';
 
 export class ChatReplayDebugSession extends LoggingDebugSession {
 
@@ -64,7 +65,7 @@ export class ChatReplayDebugSession extends LoggingDebugSession {
 				: path.join(this._workspaceFolder?.uri.fsPath || process.cwd(), programArg);
 
 			const content = fs.readFileSync(this._program, 'utf8');
-			this._chatSteps = this.parseReplay(content);
+			this._chatSteps = parseReplay(content);
 
 			this.sendResponse(response);
 
@@ -173,7 +174,6 @@ export class ChatReplayDebugSession extends LoggingDebugSession {
 		this.sendEvent(new StoppedEvent('next', ChatReplayDebugSession.THREAD_ID));
 	}
 
-
 	protected override pauseRequest(response: DebugProtocol.PauseResponse, args: DebugProtocol.PauseArguments): void {
 		// Stay on current header and report stopped
 		this.sendResponse(response);
@@ -187,75 +187,6 @@ export class ChatReplayDebugSession extends LoggingDebugSession {
 
 		this._currentIndex++;
 		return undefined;
-	}
-
-	private parsePrompt(prompt: { [key: string]: any }) {
-		const steps: ChatStep[] = [];
-		steps.push({
-			kind: 'userQuery',
-			query: prompt.prompt,
-			line: 0,
-		});
-
-		for (const log of prompt.logs) {
-			if (log.kind === 'toolCall') {
-				steps.push({
-					kind: 'toolCall',
-					id: log.id,
-					line: 0,
-					toolName: log.tool,
-					args: JSON.parse(log.args),
-					edits: log.edits,
-					results: log.response
-				});
-			} else if (log.kind === 'request') {
-				steps.push({
-					kind: 'request',
-					id: log.id,
-					line: 0,
-					prompt: log.messages,
-					result: log.response.message
-				});
-			}
-		}
-
-		return steps;
-	}
-
-	private parseReplay(content: string): ChatStep[] {
-		const parsed = JSON.parse(content);
-		const prompts = (parsed.prompts && Array.isArray(parsed.prompts) ? parsed.prompts : [parsed]) as { [key: string]: any }[];
-		if (prompts.filter(p => !p.prompt).length) {
-			throw new Error('Invalid replay content: expected a prompt object or an array of prompts in the base JSON structure.');
-		}
-
-		const steps: ChatStep[] = [];
-		for (const prompt of prompts) {
-			steps.push(...this.parsePrompt(prompt));
-		}
-
-		let stepIx = 0;
-		const lines = content.split('\n');
-		lines.forEach((line, index) => {
-			if (stepIx < steps.length) {
-				const step = steps[stepIx];
-				if (step.kind === 'userQuery') {
-					// Re-encode the query to match JSON representation in the file and remove surrounding quotes
-					const encodedQuery = JSON.stringify(step.query).slice(1, -1);
-					if (line.indexOf(`"prompt": "${encodedQuery}`) !== -1) {
-						step.line = index + 1;
-						stepIx++;
-					}
-				} else {
-					if (line.indexOf(`"id": "${step.id}"`) !== -1) {
-						step.line = index + 1;
-						stepIx++;
-					}
-				}
-
-			}
-		});
-		return steps;
 	}
 }
